feat: Enhance user experience - Optimize the clipping effect: After clipping cells, the cells produce ant lines, showing the edited effect - handleContextMenuCopy #336
