+++ src/lib/Model/PointerEventsController.ts
@@ -10,6 +10,7 @@ import { isOnClickableArea } from '../Functions/isOnClickableArea';
 import { scrollCalculator } from '../Functions/componentDidUpdate';
 import { scrollIntoView } from '../Functions/scrollIntoView';
 import { areLocationsEqual } from '../Functions/areLocationsEqual';
+import { handleCut } from '../Functions/handleCut';
 
 
 export class PointerEventsController extends AbstractPointerEventsController {
@@ -235,16 +236,7 @@ export class PointerEventsController extends AbstractPointerEventsController {
 
     private handleCut = (event: ClipboardEvent) => {
       this.updateState((state) => {
-        const elements = document.querySelectorAll(".rg-partial-area-cut-range");
-        const selectedElement = elements[0];
-        // Adds an ant line effect to the clipped cell
-        selectedElement?.classList?.add("cut-ant-line-box");
-  
-        // Set a timer to remove the effect after a certain amount of time
-        setTimeout(function () {
-          selectedElement?.classList?.remove("cut-ant-line-box");
-        }, 60000);
-  
+        handleCut();
         return state;
       });
     };
